Update API and common library components

+++ common_lib/src/models.rs
@@ -130,7 +130,7 @@ pub struct SignalWaringConfig {
     #[serde(rename = "unit")]
     pub unit: Option<String>, // 单位
     #[serde(rename = "ID")]
-    pub id: i32, // ID
+    pub id: u64, // ID
 }
 
 #[derive(Debug)]
@@ -165,7 +165,7 @@ pub struct SignalDelayWaring {
     pub name: String,
     pub script: String,
     #[serde(rename = "ID")]
-    pub id: i32,
+    pub id: u64,
 }
 
 #[derive(Debug, serde::Deserialize, serde::Serialize)]
